Дорабатывает компоненты требующие переопределения конструктора. Переносит функцию рендеринга.
