+++ src/components/task-edit.js
@@ -1,5 +1,6 @@
+import AbstractComponent from "./abstract-component.js";
 import {COLORS, DAYS, MONTHS} from "../const.js";
-import {createElement, formatTime} from "../utils.js";
+import {formatTime} from "../utils.js";
 
 const createColorsMarkup = (colors, currentColor) => {
   return colors
@@ -134,25 +135,14 @@ export const createTaskEditTemplate = (task) => {
   );
 };
 
-export default class TaskEdit {
+export default class TaskEdit extends AbstractComponent {
   constructor(task) {
+    super();
+
     this._task = task;
-    this._element = null;
   }
 
   getTemplate() {
     return createTaskEditTemplate(this._task);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
